Move Global tuning to a bundle
